water_flow_backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
python-dotenv==1.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
celery==5.5.1
redis==5.2.1
kombu==5.5.3
vine==5.1.0
billiard==4.2.1
channels==3.0.4
channels-redis==4.2.1
daphne==3.0.2
autobahn==24.4.2
twisted==24.11.0
cryptography==44.0.2
pyOpenSSL==25.0.0
service-identity==24.2.0
click==8.1.8
python-dateutil==2.9.0.post0
msgpack==1.1.0
greenlet==3.2.1
django-celery-beat==2.8.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
automat==24.8.1
cffi==1.17.1
constantly==23.10.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
six==1.17.0
txaio==23.1.1
typing-extensions==4.13.1
tzdata==2025.2
zope.interface==7.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
prompt-toolkit==3.0.51
wcwidth==0.2.13
djangorestframework_simplejwt==5.5.0
psycopg==3.2.5
PyJWT==2.9.0
pytz==2023.3.post1
psycopg2-binary==2.9.10
django-celery-results==2.6.0
celery[sqlalchemy]>=5.0